fix --open flag without a value
