+++ packages/cli/src/cli/serve.rs
@@ -4,10 +4,35 @@ use crate::{
 };
 use anyhow::Context;
 use build::Build;
-use dioxus_cli_config::ServeArguments;
+use dioxus_cli_config::AddressArguments;
 use std::ops::Deref;
 
 use super::*;
+
+/// Arguments for the serve command
+#[derive(Clone, Debug, Parser, Default)]
+pub struct ServeArguments {
+    /// The arguments for the address the server will run on
+    #[clap(flatten)]
+    pub address: AddressArguments,
+
+    /// Open the app in the default browser [default: false - unless project or global settings are set]
+    #[arg(long, default_missing_value="true", num_args=0..=1)]
+    pub open: Option<bool>,
+
+    /// Enable full hot reloading for the app [default: true - unless project or global settings are set]
+    #[clap(long, group = "release-incompatible")]
+    pub hot_reload: Option<bool>,
+
+    /// Set cross-origin-policy to same-origin [default: false]
+    #[clap(name = "cross-origin-policy")]
+    #[clap(long)]
+    pub cross_origin_policy: bool,
+
+    /// Additional arguments to pass to the executable
+    #[clap(long)]
+    pub args: Vec<String>,
+}
 
 /// Run the WASM project on dev-server
 #[derive(Clone, Debug, Default, Parser)]
